backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.19
pydantic==2.7.1
pydantic-settings==2.2.1
httpx[socks]==0.27.0
aiofiles==23.2.1
python-multipart==0.0.9
websockets==12.0
PyYAML==6.0.1
aiohttp==3.9.5
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
alembic==1.13.1
aiosqlite==0.20.0
psutil>=5.9.0
docker==7.1.0
